requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==4.2
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg==3.1.18
psycopg-binary==3.1.18
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
